fix: restore error msgs when missing db - put all in Dashboard.js
