+++ src/components/Dashboard/Dashboard.js
@@ -1,9 +1,19 @@
 import React, { Component } from 'react'
 import { connect } from 'react-redux'
 import PropTypes from 'prop-types'
+import isEmpty from 'lodash/isEmpty'
+import i18n from '@dhis2/d2-i18n'
+import { Layer, CenteredContent, CircularLoader } from '@dhis2/ui'
 
+import DashboardsBar from '../ControlBar/DashboardsBar'
+import DashboardVerticalOffset from './DashboardVerticalOffset'
+import NoContentMessage from '../../widgets/NoContentMessage'
 import { tSelectDashboard } from '../../actions/dashboards'
-import { sDashboardsIsFetching } from '../../reducers/dashboards'
+import {
+    sDashboardsIsFetching,
+    sGetAllDashboards,
+} from '../../reducers/dashboards'
+import { sGetSelectedId } from '../../reducers/selected'
 import {
     EDIT,
     NEW,
@@ -57,6 +67,42 @@ class Dashboard extends Component {
     }
 
     render() {
+        if (!this.props.dashboardsLoaded) {
+            return (
+                <Layer translucent>
+                    <CenteredContent>
+                        <CircularLoader />
+                    </CenteredContent>
+                </Layer>
+            )
+        }
+
+        if (this.props.dashboardsIsEmpty) {
+            return (
+                <>
+                    <DashboardsBar />
+                    <DashboardVerticalOffset />
+                    <NoContentMessage
+                        text={i18n.t(
+                            'No dashboards found. Use the + button to create a new dashboard.'
+                        )}
+                    />
+                </>
+            )
+        }
+
+        if (this.props.id === null) {
+            return (
+                <>
+                    <DashboardsBar />
+                    <DashboardVerticalOffset />
+                    <NoContentMessage
+                        text={i18n.t('Requested dashboard not found')}
+                    />
+                </>
+            )
+        }
+
         const ActiveDashboard = dashboardMap[this.props.mode]
 
         return <ActiveDashboard />
@@ -64,14 +110,21 @@ class Dashboard extends Component {
 }
 
 Dashboard.propTypes = {
+    dashboardsIsEmpty: PropTypes.bool,
     dashboardsLoaded: PropTypes.bool,
+    id: PropTypes.string,
     match: PropTypes.object,
     mode: PropTypes.string,
     selectDashboard: PropTypes.func,
 }
 
 const mapStateToProps = state => {
-    return { dashboardsLoaded: !sDashboardsIsFetching(state) }
+    const dashboards = sGetAllDashboards(state)
+    return {
+        dashboardsIsEmpty: isEmpty(dashboards),
+        dashboardsLoaded: !sDashboardsIsFetching(state),
+        id: sGetSelectedId(state),
+    }
 }
 
 export default connect(mapStateToProps, {
