+++ src/components/Dashboard/EditDashboard.js
@@ -8,7 +8,6 @@ import { sGetSelectedId } from '../../reducers/selected'
 import {
     sGetDashboardById,
     sGetDashboardItems,
-    sDashboardsIsFetching,
 } from '../../reducers/dashboards'
 import DashboardVerticalOffset from './DashboardVerticalOffset'
 import TitleBar from '../TitleBar/TitleBar'
@@ -16,20 +15,6 @@ import ItemGrid from '../ItemGrid/ItemGrid'
 import EditBar from '../ControlBar/EditBar'
 import NoContentMessage from '../../widgets/NoContentMessage'
 
-export const Content = ({ updateAccess }) => {
-    return updateAccess ? (
-        <>
-            <TitleBar edit={true} />
-            <ItemGrid edit={true} />
-        </>
-    ) : (
-        <NoContentMessage text={i18n.t('No access')} />
-    )
-}
-
-Content.propTypes = {
-    updateAccess: PropTypes.bool,
-}
 export class EditDashboard extends Component {
     state = {
         initialized: false,
@@ -51,26 +36,21 @@ export class EditDashboard extends Component {
             this.initEditDashboard()
         }
     }
-
-    getDashboardContent = () => {
-        const contentNotReady =
-            !this.props.dashboardsLoaded || this.props.id === null
-
-        return (
-            <div className="dashboard-wrapper">
-                {contentNotReady ? null : (
-                    <Content updateAccess={this.props.updateAccess} />
-                )}
-            </div>
-        )
-    }
-
     render() {
         return (
             <>
                 <EditBar />
                 <DashboardVerticalOffset editMode={true} />
-                {this.getDashboardContent()}
+                <div className="dashboard-wrapper">
+                    {this.props.updateAccess ? (
+                        <>
+                            <TitleBar edit={true} />
+                            <ItemGrid edit={true} />
+                        </>
+                    ) : (
+                        <NoContentMessage text={i18n.t('No access')} />
+                    )}
+                </div>
             </>
         )
     }
@@ -78,8 +58,6 @@ export class EditDashboard extends Component {
 
 EditDashboard.propTypes = {
     dashboard: PropTypes.object,
-    dashboardsLoaded: PropTypes.bool,
-    id: PropTypes.string,
     items: PropTypes.array,
     setEditDashboard: PropTypes.func,
     updateAccess: PropTypes.bool,
@@ -94,10 +72,8 @@ const mapStateToProps = state => {
 
     return {
         dashboard,
-        id,
         updateAccess,
         items: sGetDashboardItems(state),
-        dashboardsLoaded: !sDashboardsIsFetching(state),
     }
 }
