+++ src/components/Dashboard/PrintDashboard.js
@@ -1,34 +1,16 @@
 import React, { Component } from 'react'
 import { connect } from 'react-redux'
-import i18n from '@dhis2/d2-i18n'
 import PropTypes from 'prop-types'
 
 import PrintTitleBar from '../TitleBar/PrintTitleBar'
 import PrintItemGrid from '../ItemGrid/PrintItemGrid'
 import { acSetEditDashboard } from '../../actions/editDashboard'
 import { sGetSelectedId } from '../../reducers/selected'
-// import { acSetSelectedPrintMode } from '../../actions/selected'
 import {
     sGetDashboardById,
     sGetDashboardItems,
-    sDashboardsIsFetching,
 } from '../../reducers/dashboards'
-import NoContentMessage from '../../widgets/NoContentMessage'
 
-export const Content = ({ updateAccess }) => {
-    return updateAccess ? (
-        <>
-            <PrintTitleBar />
-            <PrintItemGrid />
-        </>
-    ) : (
-        <NoContentMessage text={i18n.t('No access')} />
-    )
-}
-
-Content.propTypes = {
-    updateAccess: PropTypes.bool,
-}
 export class PrintDashboard extends Component {
     state = {
         initialized: false,
@@ -37,7 +19,6 @@ export class PrintDashboard extends Component {
     initPrintDashboard = () => {
         if (this.props.dashboard) {
             this.setState({ initialized: true })
-            // this.props.setSelectedPrintMode(true)
             this.props.setEditDashboard(this.props.dashboard, this.props.items)
         }
     }
@@ -53,17 +34,10 @@ export class PrintDashboard extends Component {
     }
 
     render() {
-        const contentNotReady =
-            !this.props.dashboardsLoaded || this.props.id === null
-
         return (
             <div className="dashboard-wrapper">
-                {contentNotReady ? null : (
-                    <Content
-                        updateAccess={this.props.updateAccess}
-                        onePerPage={this.props.onePerPage}
-                    />
-                )}
+                <PrintTitleBar />
+                <PrintItemGrid />
             </div>
         )
     }
@@ -71,31 +45,21 @@ export class PrintDashboard extends Component {
 
 PrintDashboard.propTypes = {
     dashboard: PropTypes.object,
-    dashboardsLoaded: PropTypes.bool,
-    id: PropTypes.string,
     items: PropTypes.array,
-    onePerPage: PropTypes.bool,
     setEditDashboard: PropTypes.func,
-    updateAccess: PropTypes.bool,
 }
 
 const mapStateToProps = state => {
     const id = sGetSelectedId(state)
     const dashboard = id ? sGetDashboardById(state, id) : null
 
-    const updateAccess =
-        dashboard && dashboard.access ? dashboard.access.update : false
-
     return {
         dashboard,
         id,
-        updateAccess,
         items: sGetDashboardItems(state),
-        dashboardsLoaded: !sDashboardsIsFetching(state),
     }
 }
 
 export default connect(mapStateToProps, {
     setEditDashboard: acSetEditDashboard,
-    // setSelectedPrintMode: acSetSelectedPrintMode,
 })(PrintDashboard)
