+++ src/components/Dashboard/PrintLayoutDashboard.js
@@ -1,6 +1,5 @@
 import React, { Component } from 'react'
 import { connect } from 'react-redux'
-import i18n from '@dhis2/d2-i18n'
 import PropTypes from 'prop-types'
 
 import PrintTitleBar from '../TitleBar/PrintTitleBar'
@@ -13,9 +12,7 @@ import { sGetSelectedId } from '../../reducers/selected'
 import {
     sGetDashboardById,
     sGetDashboardItems,
-    sDashboardsIsFetching,
 } from '../../reducers/dashboards'
-import NoContentMessage from '../../widgets/NoContentMessage'
 import { getNumPrintPages } from '../../modules/printUtils'
 import { PAGEBREAK } from '../../modules/itemTypes'
 
@@ -31,20 +28,6 @@ const addPageBreaks = ({ items, addDashboardItem }) => {
     return items
 }
 
-export const Content = ({ updateAccess }) => {
-    return updateAccess ? (
-        <>
-            <PrintTitleBar />
-            <PrintLayoutItemGrid />
-        </>
-    ) : (
-        <NoContentMessage text={i18n.t('No access')} />
-    )
-}
-
-Content.propTypes = {
-    updateAccess: PropTypes.bool,
-}
 export class PrintLayoutDashboard extends Component {
     state = {
         initialized: false,
@@ -72,14 +55,10 @@ export class PrintLayoutDashboard extends Component {
     }
 
     render() {
-        const contentNotReady =
-            !this.props.dashboardsLoaded || this.props.id === null
-
         return (
             <div className="dashboard-wrapper">
-                {contentNotReady ? null : (
-                    <Content updateAccess={this.props.updateAccess} />
-                )}
+                <PrintTitleBar />
+                <PrintLayoutItemGrid />
             </div>
         )
     }
@@ -87,26 +66,18 @@ export class PrintLayoutDashboard extends Component {
 
 PrintLayoutDashboard.propTypes = {
     dashboard: PropTypes.object,
-    dashboardsLoaded: PropTypes.bool,
-    id: PropTypes.string,
     items: PropTypes.array,
     setEditDashboard: PropTypes.func,
-    updateAccess: PropTypes.bool,
 }
 
 const mapStateToProps = state => {
     const id = sGetSelectedId(state)
     const dashboard = id ? sGetDashboardById(state, id) : null
 
-    const updateAccess =
-        dashboard && dashboard.access ? dashboard.access.update : false
-
     return {
         dashboard,
         id,
-        updateAccess,
         items: sGetDashboardItems(state),
-        dashboardsLoaded: !sDashboardsIsFetching(state),
     }
 }
