+++ src/components/Dashboard/ViewDashboard.js
@@ -1,57 +1,16 @@
 import React, { useEffect } from 'react'
 import { connect } from 'react-redux'
-import isEmpty from 'lodash/isEmpty'
-import i18n from '@dhis2/d2-i18n'
 import PropTypes from 'prop-types'
 
-import {
-    sGetAllDashboards,
-    sDashboardsIsFetching,
-} from '../../reducers/dashboards'
-import { sGetIsEditing } from '../../reducers/editDashboard'
-import { sGetSelectedId } from '../../reducers/selected'
-import { acClearEditDashboard } from '../../actions/editDashboard'
-import DashboardsBar from '../ControlBar/DashboardsBar'
-import DashboardVerticalOffset from './DashboardVerticalOffset'
 import TitleBar from '../TitleBar/TitleBar'
 import ItemGrid from '../ItemGrid/ItemGrid'
 import FilterBar from '../FilterBar/FilterBar'
+import DashboardsBar from '../ControlBar/DashboardsBar'
+import DashboardVerticalOffset from './DashboardVerticalOffset'
+import { sGetIsEditing } from '../../reducers/editDashboard'
+import { acClearEditDashboard } from '../../actions/editDashboard'
 
-import NoContentMessage from '../../widgets/NoContentMessage'
-
-export const Content = ({ hasDashboardContent, dashboardsIsEmpty }) => {
-    const msg = dashboardsIsEmpty
-        ? i18n.t(
-              'No dashboards found. Use the + button to create a new dashboard.'
-          )
-        : i18n.t('Requested dashboard not found')
-
-    return hasDashboardContent ? (
-        <>
-            <TitleBar edit={false} />
-            <FilterBar />
-            <ItemGrid edit={false} />
-        </>
-    ) : (
-        <NoContentMessage text={msg} />
-    )
-}
-
-Content.propTypes = {
-    dashboardsIsEmpty: PropTypes.bool,
-    hasDashboardContent: PropTypes.bool,
-}
-
-export const ViewDashboard = ({
-    id,
-    dashboardsIsEmpty,
-    dashboardsLoaded,
-    dashboardIsEditing,
-    clearEditDashboard,
-}) => {
-    const hasDashboardContent = id && !dashboardsIsEmpty
-    const contentNotReady = !dashboardsLoaded || id === null
-
+export const ViewDashboard = ({ dashboardIsEditing, clearEditDashboard }) => {
     useEffect(() => {
         if (dashboardIsEditing) {
             clearEditDashboard()
@@ -63,12 +22,9 @@ export const ViewDashboard = ({
             <DashboardsBar />
             <DashboardVerticalOffset />
             <div className="dashboard-wrapper">
-                {contentNotReady ? null : (
-                    <Content
-                        hasDashboardContent={hasDashboardContent}
-                        dashboardsIsEmpty={dashboardsIsEmpty}
-                    />
-                )}
+                <TitleBar edit={false} />
+                <FilterBar />
+                <ItemGrid edit={false} />
             </div>
         </>
     )
@@ -77,18 +33,10 @@ export const ViewDashboard = ({
 ViewDashboard.propTypes = {
     clearEditDashboard: PropTypes.func,
     dashboardIsEditing: PropTypes.bool,
-    dashboardsIsEmpty: PropTypes.bool,
-    dashboardsLoaded: PropTypes.bool,
-    id: PropTypes.string,
 }
 
 const mapStateToProps = state => {
-    const dashboards = sGetAllDashboards(state)
-
     return {
-        id: sGetSelectedId(state),
-        dashboardsIsEmpty: isEmpty(dashboards),
-        dashboardsLoaded: !sDashboardsIsFetching(state),
         dashboardIsEditing: sGetIsEditing(state),
     }
 }
